Fixed all of the problems so that when you click on the Notification that's generated when entering or leaving a fence it correctly displays the information about your trip.  This definitely needs some refinement to display more information as well as giving the user some options about what to do with the data.  I also modified the SQLite tables to better reflect what I care about.  It now save all trip segments regardless of whether or not they were business related so I can use them for distance lookups.  Not only does it save them but there is now a boolean trigger that lets the program know whether or not they are business related.
I created a new table that holds information about Trip groupings.  This table tracks when new trips are started as well as when they're over.  It has two columns, one as a unique identifier for each trip group and another detailing when it's closed.  The one that details when it is closed is accessed by a foreign key in the TripTable so that when you close a group it automatically closes all related path segments.

There is a new class that handles creating Trip Rows.  It automatically handles whether or not your trip is closed as well as handling inserting new data.