Starting at the top....
ExpandingTripList:

I fixed a bug where sometimes laoding the data would cause an exception.  I fixed it by creating an object to store the list items in until they're all loaded.  It then shallow copies them all into the list and notifies the adapter.

ShowLocation:
This is a new class that displays a Google Map.  At the moment it needs a Bundles containing a serialized ExpandListChild that is the origin of the click event that creates this.  It then takes that ExpandListChild and gets the ExpandListGroup that contains it.  It takes that group and gets the points associated with each child.  It takes those points and makes a call to the Google Servers to pull down all the route point information for each trip point.  It then takes all of those points and creates Polylines for them to display on the map.

ExpandListChild:
Added a way to store all of the route poitns and a reference to its container object

ExpandListGroup:
Minor methods to get the count of children and the container of all the children

AddressDistanceServices:

Added new methods to create a URL to get all the address points.  Also added helper methods to convert those points into objects that I can use.