STILL BROKEN, DO NOT USE!
I went through the date/time handling and modified the code so that it correctly records when your stops happened.  Meaning that when you start driving again it records that time.  When you stop and record a TripRow it then uses that start time with the current time to record that row.  That makes it possible to reference those times and calculate how long a trip was.