STILL BROKEN, DO NOT USE!
Finished the GetCurrentLocation class.  It now handles all the location needs of the app.  It will create geofences, automatically select the best location provider and resolution, and return the best location that it can get.  All classes that were using individual location providers have been migrated to the new provider.