STILL NOT FULLY FUNCTIONAL!
So, I created a new interface in the MapDrawerActivity class so that I can manage connections to the Location services when the status of the app changes.  This means that when the app is paused it now disconnects everything.  It reconnects everything if you unPause the app.

I also ported all of the code to create HomePoints and geofences into the new HomeHandler class.  It now can perform CRUD operations on the database to manage HomePoints as well as create and remove the geofences.  This portion of the code is largely complete, I'm sure there are outstanding bugs that I need to find but for now everything seems to be working as intended.  There is now a lot of testing that needs to happen in order to ensure that trips are being recorded properly.