The slide up drawer on the bottom now works.  I created a special adapter for it(there's still a bug it where the data is not resetting correctly.) It also has its own layouts and XML definitions to keep it looking pretty.
I modified the WifiReceiver and the CalculateMileage classes.  I moved all of the functionality for the network checking and calculation into the CalculateMileage IntentService.  This serves two purposes, if the network connection is acting wonky it won't cause the BroadcastReceiver to time out and be killed causing all kinds of havoc and it moves all of the checking and calculation to a background thread.