Whew, where to begin.  I"m working on getting all the callbacks working.  Making the map system work so that when you click on an item it draws the route on the map.
I created a class that will process groups the same way every time.  While it's a fairly complicated class, it does manage all aspects ofTripGroups.  From verifying their validity to updating all of the row information and marking the TripGroup as processed.

I also added a couple different columns to the database.  The TripGroup now has a processed INTEGER column that tracks whether or not the TripGroup has been processed.  The TripRow now has a points TEXT column that will store the encoded polyline.  This will keep the app from having to download that every time.
Next up are to get the PolyLines actually drawing and to add the ability to move/delete stops.